Only need long? metrics in IStats.Event()

+++ Hudl.Mjolnir/External/IStats.cs
@@ -4,7 +4,7 @@ namespace Hudl.Mjolnir.External
 {
     public interface IStats
     {
-        void Event(string service, string state, float? metric);
+        void Event(string service, string state, long? metric);
         void Elapsed(string service, string state, TimeSpan elapsed);
         void Gauge(string service, string state, long? metric = null);
     }
